pyproject.toml
[project]
name = "tradefun"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolai",email = "vlasov.bis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "t-tech-investments (>=0.3.3,<0.4.0)",
    "dependency-injector (>=4.48.3,<5.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
